Statistics of chemical shifts values of beta carbons from RefDB.

A dataset containing statistics of chemical shifts of beta carbons for each amino acid from RefDB data:

Residue. amino acid typing (single-letter convention)

Coil.mu. mean value of beta carbon in coil

Coil.sd. mean value of beta carbon in coil

Helix.mu. mean value of beta carbon in helix

Helix.sd. standard deviation value of beta carbon in helix

Beta.mu. mean value of beta carbon in coil

Beta.sd. standard deviation value of beta carbon in coil

Avg.mu. average of mean values of beta carbon across three secondary structures

Avg.sd. average of standard deviation values of beta carbon across three secondary structures

RefDB.StatCB

Format

A data frame with 21 rows and 9 columns.