RefDB Data Sets (R6 class object).

Collection of data from Re-referenced Protein Chemical shift Database (RefDB). Each data entry in the database represent one protein and the corresponding chemical shifts from NMR experiment for that protein. The original data are from the BioMagRes Bank, and all the data here are carefully corrected or re-referenced using X-ray or NMR coordinate data via SHIFTX and then comparing those predictions to the observed shifts reported in the BMRB (via SHIFTCOR) [1].

RefDB_data

Format

A data frame with 1557 13C data sets:

entry

ID, BMRB ID

table

table contains ID, AA, SS, CA, CB, which represent ID: Residue Potition of Protein, AA: Amino Acid Typing, SS: Secondary Structure, CA: Alpha Carbon Chemical Shifts, CB: Beta Carbon Chemical Shifts

usage

RefDB_data$

Source

http://refdb.wishartlab.com/

Details

Currently there are 2162 entry in the RefDB, but we only use 1557 of them which contain carbon chemical shifts.

Reference:

[1] Haiyan Zhang, Stephen Neal and David Wishart (2003) "RefDB: A database of uniformly referenced protein chemical shifts" Journal of Biomolecular NMR, 25: 173-195