calculateChiSquaredStat

calculateChiSquaredStat Given a pair of C_alpha and C_beta chemical shifts, this function will return a list of caculated chisquare statistics basing on the combination of amino acid typings and secondary structures.

calculateChiSquaredStat(cacbPair)

Arguments

x
A pair of carbon chemical shifts (Ca, Cb)

Value

A list of chisquare statisitcs basing the combination of amino acid typings and secondary structures.

Examples

calculateChiSquaredStat(c(54,45))

#>     AA                H                 B                C
#>  1:  A 1023.16868895477  169.099242775473 469.048336795045
#>  2:  B 180.178895468627  54.4937018085712 74.2438106347521
#>  3:  C 24.0426461856623  6.02618817827703 23.2477754200126
#>  4:  D 14.4848786940741  2.79850128029531 10.2715883462805
#>  5:  E 13.3897615181103  1.88818580572447 4.60208889084118
#>  6:  F 165.514869003217  77.9438324817418 62.5815610717518
#>  7:  H 321.477487110107  53.4257171238817 91.6072511229365
#>  8:  I 253.099344063356  44.4933903131377 94.8605497904786
#>  9:  K 128.989994525817  35.9002492383691 39.0974374716513
#> 10:  L 100.123349476134  19.1632752720135 22.6939781262421
#> 11:  M 14.8396700069463 0.383003397426531 2.60892370474794
#> 12:  N 222.558693666307  34.5192736984713 53.7891077582836
#> 13:  P 89.4591494720313    19.26900894692 26.5279676961367
#> 14:  Q 40.3446974047416  5.81357432329253 11.1635747587969
#> 15:  R 267.220874922648  179.412483703255 236.676686103036
#> 16:  S 203.258956942293  177.779740165393 190.175848300641
#> 17:  T 451.922344250779  363.817995197962 408.684228724244
#> 18:  V 58.3956321465275  6.77558973670799 13.8132678105348
#> 19:  W 127.135809610673  63.1703556188059 77.6629767003917
#> 20:  Y 358.237467618581  47.5540232633832 80.5896644421314