read_DB_file() reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.

read_DB_file() reads in data from existing database that included in the BaMORC package. This database was extracted from RefDB database.

read_DB_File(id)

Arguments

id
BMRB or RefDB entry ID.

Value

Protein sequence, secondary structure information and chemical shifts dataframe.

Examples

id = 4022
head(read_DB_File(id))

#> [[1]]
#>   [1] "A" "S" "P" "D" "W" "G" "Y" "D" "D" "K" "N" "G" "P" "E" "Q" "W" "S" "K"
#>  [19] "L" "Y" "P" "I" "A" "N" "G" "N" "N" "Q" "S" "P" "V" "D" "I" "K" "T" "S"
#>  [37] "E" "T" "K" "H" "D" "T" "S" "L" "K" "P" "I" "S" "V" "S" "Y" "N" "P" "A"
#>  [55] "T" "A" "K" "E" "I" "I" "N" "V" "G" "H" "S" "F" "H" "V" "N" "F" "E" "D"
#>  [73] "N" "D" "N" "R" "S" "V" "L" "K" "G" "G" "P" "F" "S" "D" "S" "Y" "R" "L"
#>  [91] "F" "Q" "F" "H" "F" "H" "W" "G" "S" "T" "N" "E" "H" "G" "S" "E" "H" "T"
#> [109] "V" "D" "G" "V" "K" "Y" "S" "A" "E" "L" "H" "V" "A" "H" "W" "N" "S" "A"
#> [127] "K" "Y" "S" "S" "L" "A" "E" "A" "A" "S" "K" "A" "D" "G" "L" "A" "V" "I"
#> [145] "G" "V" "L" "M" "K" "V" "G" "E" "A" "N" "P" "K" "L" "Q" "K" "V" "L" "D"
#> [163] "A" "L" "Q" "A" "I" "K" "T" "K" "G" "K" "R" "A" "P" "F" "T" "N" "F" "D"
#> [181] "P" "S" "T" "L" "L" "P" "S" "S" "L" "D" "F" "W" "T" "Y" "P" "G" "S" "L"
#> [199] "T" "H" "P" "L" "Y" "E" "S" "V" "T" "W" "I" "I" "C" "K" "E" "S" "I" "S"
#> [217] "V" "S" "S" "E" "Q" "L" "A" "Q" "F" "R" "S" "L" "L" "S" "N" "V" "E" "G"
#> [235] "D" "N" "A" "V" "P" "M" "Q" "H" "N" "N" "R" "P" "T" "Q" "P" "L" "K" "G"
#> [253] "R" "T" "V" "R" "A" "S" "F"
#> 
#> [[2]]
#>       CA   CB
#> 1   51.9 19.4
#> 112 57.5 63.3
#> 183 63.1 32.2
#> 194 54.6 41.3
#> 205 54.7 31.4
#> 216 45.4   NA
#> 227 58.6 39.7
#> 238 53.9 42.2
#> 249 56.9 40.5
#> 2   57.3 32.7
#> 13  52.2 38.8
#> 24  45.5   NA
#> 35  64.2 32.0
#> 46  58.5 28.7
#> 57  55.1 29.1
#> 68  58.5 29.2
#> 79  60.2 62.0
#> 90  58.6 32.3
#> 101 55.2 44.8
#> 113 55.0 39.0
#> 123 65.2 31.8
#> 134 60.7 38.9
#> 145 53.7 18.6
#> 156 52.2 38.5
#> 167 45.0   NA
#> 178 54.1 39.4
#> 180 51.4 39.8
#> 181 55.5 31.9
#> 182 57.0 63.2
#> 184 61.8 33.4
#> 185 59.0 35.5
#> 186 52.4 42.3
#> 187 61.6 37.6
#> 188 54.2 31.7
#> 189 66.2 68.9
#> 190 59.9 62.8
#> 191 55.4 30.8
#> 192 61.4 71.1
#> 193 54.0 34.6
#> 195 55.3 27.9
#> 196 52.5 42.0
#> 197 63.1 68.4
#> 198 59.4 63.7
#> 199 55.1 42.0
#> 200 53.6 31.8
#> 201 63.0 31.8
#> 202 60.5 38.8
#> 203 56.5 64.4
#> 204 59.2 33.1
#> 206 56.0 62.8
#> 207 52.4 38.0
#> 208 49.2 41.0
#> 209 64.4 32.0
#> 210 53.8 18.3
#> 211 62.8 70.7
#> 212 55.0 17.9
#> 213 57.7 37.0
#> 214 55.4 33.6
#> 215 58.3 42.3
#> 217 59.2 43.0
#> 218 53.1 38.2
#> 219 59.8 30.2
#> 220 47.1   NA
#> 221 54.6 31.0
#> 222 56.3 62.8
#> 223 55.8 42.4
#> 224 52.2 35.2
#> 225 61.4 30.4
#> 226 52.3 41.5
#> 228 56.4 41.8
#> 229 57.7 30.8
#> 230 53.5 40.7
#> 231 54.1 39.1
#> 232 52.8 41.5
#> 233 52.1 37.5
#> 234 60.4 32.1
#> 235 55.9 65.6
#> 236 58.9 37.7
#> 237 53.4 45.7
#> 239 54.6 36.6
#> 240 43.9   NA
#> 241 44.6   NA
#> 242 62.7 32.1
#> 243 57.6 39.9
#> 244 59.1 64.0
#> 245 53.4 44.0
#> 246 57.7 64.5
#> 247 56.2 41.4
#> 248 55.5 32.5
#> 250 56.2 41.9
#> 251 57.8 41.9
#> 252 56.0 35.8
#> 253 57.1 44.0
#> 254 55.3 32.0
#> 255 55.9 42.8
#> 256 53.7 32.3
#> 257 57.4 30.6
#> 258 43.0   NA
#> 259 61.1 63.4
#> 3   58.8 71.6
#> 4   54.8 37.5
#> 5   58.5 29.7
#> 6   53.9 31.6
#> 7   46.7   NA
#> 8   58.2 63.7
#> 9   58.6 30.1
#> 10  60.3 28.6
#> 11  59.6 71.1
#> 12  61.2 34.2
#> 14  56.5 39.3
#> 15  45.0   NA
#> 16  63.5 31.9
#> 17  55.9 33.5
#> 18  57.5 38.5
#> 19  62.5 64.0
#> 20  52.0 22.5
#> 21  56.2 36.0
#> 22  53.8 43.9
#> 23  52.5 33.6
#> 25  61.8 32.4
#> 26  52.6 23.0
#> 27  54.8 34.3
#> 28  54.8 34.3
#> 29  53.4 37.4
#> 30  57.9 62.5
#> 31  53.4 19.0
#> 32  55.9 34.6
#> 33  56.0 40.6
#> 34  59.2 64.6
#> 36  56.8 65.5
#> 37  56.9 39.9
#> 38  54.5 17.8
#> 39  58.2 30.8
#> 40  54.3 20.3
#> 41  55.6 17.8
#> 42  56.5 63.3
#> 43  53.5 33.3
#> 44  55.5 18.8
#> 45  52.4 39.5
#> 47  48.7   NA
#> 48  53.4 46.0
#> 49  50.8 23.2
#> 50  59.7 33.1
#> 51  59.7 40.1
#> 52  43.9   NA
#> 53  60.2 34.6
#> 54  55.3 41.1
#> 55  54.0 34.7
#> 56  54.0 36.5
#> 58  61.5 31.2
#> 59  46.7   NA
#> 60  56.3 30.1
#> 61  52.8 18.4
#> 62  52.8 39.8
#> 63  64.4 32.2
#> 64  57.4 31.3
#> 65  54.6 43.4
#> 66  59.9 28.9
#> 67  59.9 33.0
#> 69  65.6 31.9
#> 70  57.3 40.1
#> 71  56.8 40.5
#> 72  53.2 19.6
#> 73  56.8 40.0
#> 74  57.6 27.8
#> 75  52.5 20.1
#> 76  58.7 38.4
#> 77  60.4 33.7
#> 78  58.3 71.0
#> 80  58.3 32.0
#> 81  44.3   NA
#> 82  56.6 33.2
#> 83  53.6 34.6
#> 84  49.8 20.8
#> 85  63.0 32.0
#> 86  58.4 40.6
#> 87  58.6 72.0
#> 88  53.7 37.0
#> 89  58.0 44.1
#> 91  48.8 42.2
#> 92  64.0 31.3
#> 93  61.8   NA
#> 94  63.7 70.0
#> 95  54.4 42.9
#> 96  52.9 39.7
#> 97  62.2 32.2
#> 98  60.8 63.1
#> 99  56.3 64.2
#> 100 53.6 40.5
#> 102 56.2 40.4
#> 103 56.8 43.0
#> 104 57.3 31.5
#> 105 58.2 70.6
#> 106 55.9 39.9
#> 107 62.0 31.4
#> 108 46.3   NA
#> 109 56.7 70.8
#> 110 55.0 43.7
#> 111 61.5 67.8
#> 114 51.9 28.8
#> 115 64.1 31.6
#> 116 57.1 37.4
#> 117 56.2 38.7
#> 118 55.7 26.3
#> 119 56.0 66.9
#> 120 61.7 35.6
#> 121 63.3 68.2
#> 122 58.2 31.6
#> 124 58.2 38.2
#> 125 56.9 37.1
#> 126 62.6 27.1
#> 127 57.8 33.0
#> 128 55.4 30.3
#> 129 56.9 66.3
#> 130 60.0 40.9
#> 131 57.6 65.9
#> 132 60.5 33.7
#> 133 56.4 66.1
#> 135 56.4 66.0
#> 136 59.9 28.7
#> 137 59.4 28.7
#> 138 57.6 41.5
#> 139 54.5 17.8
#> 140 59.0 28.2
#> 141 57.8 34.1
#> 142 56.5 30.3
#> 143 58.9 63.8
#> 144 55.5 42.8
#> 146 53.9 41.4
#> 147 58.6 66.1
#> 148 51.9 42.1
#> 149 61.5 33.1
#> 150 58.5 29.5
#> 151 45.1   NA
#> 152 52.6 42.5
#> 153 54.0 38.5
#> 154 52.7 19.4
#> 155 59.5 34.6
#> 157 63.4 32.8
#> 158 54.9 32.5
#> 159 59.6 30.7
#> 160 53.8 31.2
#> 161 51.7 36.4
#> 162 51.8 39.0
#> 163 53.9   NA
#> 164 62.6 33.0
#> 165 62.5 70.2
#> 166 51.4 29.2
#> 168 62.3 31.8
#> 169 57.9 42.3
#> 170 56.9 29.7
#> 171 45.5   NA
#> 172 56.9 33.6
#> 173 62.2 70.8
#> 174 61.7 31.5
#> 175 54.4 33.9
#> 176 49.7 23.0
#> 177 58.6 63.6
#> 179 55.8 40.7
#>